+++ src/main/java/net/rodofire/mushrooomsmod/world/mushrooms/codemushrooms/CustomGreenMushroom.java
@@ -1,4 +1,4 @@
-package net.rodofire.mushrooomsmod.world.tree.mushrooms.codemushrooms;
+package net.rodofire.mushrooomsmod.world.mushrooms.codemushrooms;
 
 import com.mojang.serialization.Codec;
 import net.rodofire.mushrooomsmod.feature.mushroomfeature.ModMushroomFeatureConfig;
