+++ src/main/java/net/rodofire/mushrooomsmod/world/mushrooms/structuremushrooms/CustomGreenSecondMushroom.java
@@ -1,4 +1,4 @@
-package net.rodofire.mushrooomsmod.world.tree.mushrooms.structuremushrooms;
+package net.rodofire.mushrooomsmod.world.mushrooms.structuremushrooms;
 
 import com.mojang.serialization.Codec;
 import net.minecraft.block.BlockState;
