+++ src/main/java/net/rodofire/mushrooomsmod/world/mushrooms/structuremushrooms/CustomGreenSecondMushroomFeature.java
@@ -1,4 +1,4 @@
-package net.rodofire.mushrooomsmod.world.tree.mushrooms.structuremushrooms;
+package net.rodofire.mushrooomsmod.world.mushrooms.structuremushrooms;
 
 import com.mojang.serialization.Codec;
 import me.emafire003.dev.structureplacerapi.StructurePlacerAPI;
